requirements.txt
fastapi~=0.115.6
pydantic~=2.10.5
supabase~=2.11.0
python-dotenv~=1.0.1
requests~=2.32.3
openai~=1.59.8
numpy~=2.2.2
matplotlib~=3.10.0